% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLZ_prior.R
\docType{class}
\name{MLZ_prior-class}
\alias{MLZ_prior-class}
\alias{MLZ_prior}
\title{Class-\code{MLZ_prior}}
\description{
An S4 class that contains priors and model structure input. See example for template code.
}
\section{Slots}{

\describe{
\item{\code{ncp}}{The number of change points in mortality.}

\item{\code{Z_dist}}{Character. The probability distribution for the mortality rate. Currently supports \code{lognormal} or \code{uniform}.
Defaults to \code{uniform}.}

\item{\code{Z_par}}{A matrix with \code{ncp+1} rows and 2 columns. Parameters describing the priors for the mortality rate. 
If lognormal, the first column is the mean and the second column is the standard deviation of the prior (in normal space).
If uniform, the first column is the minimum and the second column is the maximum of the distribution. Defaults to \code{uniform} with bounds 0 and 3.}

\item{\code{alpha}}{The alpha vector of the Dirichlet distribution of length \code{ncp+1} for parameterization the priors for the change points 
(when \code{ncp > 0}). Ignored if \code{ncp = 0}. Defaults to \code{rep(1, ncp + 1)} which is a relatively uninformative prior.}

\item{\code{sigma_dist}}{The probability distribution for the standard deviation. Currently supports \code{lognormal} or \code{uniform}.
Defaults to \code{uniform}.}

\item{\code{sigma_par}}{A vector of length 2. Parameters describing the priors for the standard deviation. 
If lognormal, the mean and standard deviation of the prior (in normal space).
If uniform, the first column is the minimum and the second column is the maximum of the distribution. Defaults to \code{uniform} with bounds 0 and 1e5.}
}}

\author{
Q. Huynh
}
