% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit_methods.R
\docType{methods}
\name{plot,stanfit,missing-method}
\alias{plot,stanfit,missing-method}
\title{Plot function for stanfit objects in MLZBayes}
\usage{
\S4method{plot}{stanfit,missing}(x, stan_prior = NULL, subplot = c("par",
  "ts"), interval = 0.95, ...)
}
\arguments{
\item{x}{A stanfit object from \link{ML_stan}.}

\item{stan_prior}{An optional stanfit object for the same model as \code{x} but only sampled the priors.}

\item{subplot}{A vector of plots to draw (ts = time series, par = parameter posteriors)}

\item{interval}{The probability coverage interval for time series plots.}
}
\description{
Plot function for stanfit objects in MLZBayes
}
