% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit_methods.R
\docType{methods}
\name{plot,stanfit,missing-method}
\alias{plot,stanfit,missing-method}
\alias{plot.MLZBayes}
\title{Plot function for stanfit objects in MLZBayes}
\usage{
\S4method{plot}{stanfit,missing}(x, subplot = c("par", "ts"),
  interval = 0.95)

\S4method{plot}{stanfit,stanfit}(x, y, subplot = c("par", "ts"),
  interval = 0.95)
}
\arguments{
\item{x}{A stanfit object from \link{ML_stan}.}

\item{subplot}{A character vector of plots to draw. (\code{"par"} for parameter priors/posteriors, 
\code{"ts"} for time series of mean length and Z.}

\item{interval}{The probability coverage interval for the time series plots.}

\item{y}{An optional stanfit object for the same model as \code{x} but only sampled the priors. See example.}
}
\description{
This function produces up two plots. The first plot draws density plots for priors and posteriors
of model parameters (dotted lines indicate median). The second plot is a time series plot of observed (black) and predicted 
(red, with median in bolded lines and confidence intervals in dotted lines) mean lengths and estimated Z.
}
\examples{
\dontrun{
library(MLZ)
data(Goosefish)

# Create an object with priors for a model with 2 change points in mortality (ncp = 2)
uninformative_priors <- new("MLZ_prior", ncp = 2)

# Run the MCMC (calls rstan::sampling)
res <- ML_stan(Goosefish, uninformative_priors)
plot(res)

# Test priors only
res2 <- ML_stan(Goosefish, uninformative, prior_only = TRUE)
plot(res2)

# Overlay posteriors on top of priors
plot(res, res2) 
}
}
\seealso{
\link{ML_stan}
}
