% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLZprior.R
\docType{class}
\name{MLZ_prior-class}
\alias{MLZ_prior-class}
\alias{MLZ_prior}
\title{Class-\code{MLZ_prior}}
\description{
An S4 class that contains priors and model structure input.
}
\section{Slots}{

\describe{
\item{\code{ncp}}{The number of change points in mortality.}

\item{\code{Z_dist}}{The probability distribution for the mortality rate. Currently supports lognormal.}

\item{\code{Z_par}}{Parameters describing the priors for the mortality rate.}

\item{\code{cp_par}}{The alpha for the change points (uses Dirichlet formulation, see vignette).}

\item{\code{sigma_dist}}{The probability distribution for the standard deviation. Currently supports lognormal.}

\item{\code{sigma_par}}{Parameters describing the priors for the standard deviation.}
}}

\author{
Q. Huynh
}
